"pre-flatten STDIN before passing to functions"
