+++ ergonomica/lib/lib/ergo_write.py
@@ -8,24 +8,7 @@ Defines the "write" command.
 """
 
 
-import types
 
-def flatten(S):
-    if S == []:
-        return S
-    if isinstance(S[0], list):
-        return flatten(S[0]) + flatten(S[1:])
-    return S[:1] + flatten(S[1:])
-
-def flatten_stdin(stdin):
-    if isinstance(stdin, types.GeneratorType):
-        _sum = []
-        for i in stdin:
-            _sum += flatten_stdin(i)
-        return _sum
-    else:
-        return stdin
-        
 
 def main(argc):
     """write: Write STDIN to file FILE.
@@ -33,5 +16,5 @@ def main(argc):
     Usage:
         write <file>FILE
     """
-    
-    open(argc.args['FILE'], 'w').write('\n'.join(flatten([flatten_stdin(x) for x in argc.stdin])))
+
+    open(argc.args['FILE'], 'w').write('\n'.join(argc.stdin))
